Use transactions in graph storage. Fix some bugs with graph storage operations.

+++ ftd-backend/ftd-types/src/graph.rs
@@ -10,6 +10,6 @@ pub struct GraphUpdaterState {
 pub struct TransferVolume {
     pub from: String,
     pub to: String,
-    pub count: u128,
-    pub volume: u32,
+    pub count: u32,
+    pub volume: u128,
 }
